* Some refactoring to use TimeInterval wherever appropriate (durations, etc), and Double instead of Float for most other things. * ActionTween continues to tween a Float value because RealityKit tends to use Float for most of its components within Transforms, SIMD3<Float>, etc. * For the RotateBy/RotateTo I've added overloads of the initialiser that allow angles in Radians to be provided.  This meant some internal changes so that Radians are used internally.  I've deliberately kept the original initialiser as accepting angles in Degrees so that it is backwards compatible. * Added a simple test to verify that the alterations I did to the QuatMath hadn't broken anything.
It seems like a lot of change, but its a fairly simple quality of life change in that the repeated casting between TimeInterval and Float to use this wonderful Package is eliminated without any real impact to its function.

+++ Sources/RealityActions/Ease/EaseBackOut.swift
@@ -28,7 +28,7 @@ class EaseBackOutState: ActionEaseState {
         super.init(action: action, target: target)
     }
     
-    override func update(time: Float) {
+    override func update(time: Double) {
         innerActionState.update (time: backOut (time: time))
     }
 }
