+++ Sources/RealityActions/Ease/EaseExponentialOut.swift
@@ -28,7 +28,7 @@ class EaseExponentialOutState: ActionEaseState {
         super.init(action: action, target: target)
     }
     
-    override func update(time: Float) {
+    override func update(time: Double) {
         innerActionState.update (time: exponentialOut (time: time))
     }
 }
