+++ Sources/RealityActions/Ease/EaseInOut.swift
@@ -10,7 +10,7 @@ import RealityKit
 
 /// Easing function: ease-in-and-then-out
 public class EaseInOut: EaseRateAction {
-    public override init (_ action: FiniteTimeAction, rate: Float) {
+    public override init (_ action: FiniteTimeAction, rate: Double) {
         super.init(action, rate: rate)
     }
     
@@ -28,12 +28,12 @@ class EaseInOutState: EaseRateActionState {
         super.init(action: action, target: target)
     }
     
-    override func update(time: Float) {
+    override func update(time: Double) {
         let atime = time * 2
         if atime < 2 {
-            innerActionState.update (time: 0.5 * powf (atime, rate))
+            innerActionState.update (time: 0.5 * pow (atime, rate))
         } else {
-            innerActionState.update (time: 1 - 0.5 * powf (2-atime, rate))
+            innerActionState.update (time: 1 - 0.5 * pow (2-atime, rate))
         }
     }
 }
