+++ Sources/RealityActions/EaseMath.swift
@@ -7,21 +7,21 @@
 
 import Foundation
 
-func backIn(time: Float) -> Float {
-    let overshoot: Float = 1.70158
+func backIn(time: Double) -> Double {
+    let overshoot: Double = 1.70158
     
     return time * time * ((overshoot + 1) * time - overshoot)
 }
 
-func backOut(time: Float) -> Float {
-    let overshoot: Float = 1.70158
+func backOut(time: Double) -> Double {
+    let overshoot: Double = 1.70158
     
     let time = time - 1
     return time * time * ((overshoot + 1) * time + overshoot) + 1
 }
 
-func backInOut(time: Float) -> Float {
-    let overshoot: Float = 1.70158 * 1.525
+func backInOut(time: Double) -> Double {
+    let overshoot: Double = 1.70158 * 1.525
     
     let time = time * 2
     if time < 1 {
@@ -32,7 +32,7 @@ func backInOut(time: Float) -> Float {
     }
 }
 
-func bounceOut(time: Float) -> Float {
+func bounceOut(time: Double) -> Double {
     if time < 1 / 2.75 {
         return 7.5625 * time * time
     } else if (time < 2 / 2.75) {
@@ -47,11 +47,11 @@ func bounceOut(time: Float) -> Float {
     return 7.5625 * utime * utime + 0.984375
 }
 
-func bounceIn(time: Float) -> Float {
+func bounceIn(time: Double) -> Double {
     return 1 - bounceOut(time: 1 - time)
 }
 
-func bounceInOut(time: Float) -> Float {
+func bounceInOut(time: Double) -> Double {
     if time < 0.5 {
         let time = time * 2
         return (1 - bounceOut(time: 1 - time)) * 0.5
@@ -59,55 +59,55 @@ func bounceInOut(time: Float) -> Float {
     return bounceOut(time: time * 2 - 1) * 0.5 + 0.5
 }
 
-func sineOut(time: Float) -> Float {
+func sineOut(time: Double) -> Double {
     sin (time * .pi/2)
 }
 
-func sineIn(time: Float) -> Float {
+func sineIn(time: Double) -> Double {
     -1 * cos(time * .pi/2) + 1
 }
 
-func sineInOut(time: Float) -> Float {
+func sineInOut(time: Double) -> Double {
     -0.5 * (cos(.pi * time) - 1)
 }
 
-func exponentialOut(time: Float) -> Float {
-    time == 1 ? 1 : (-powf(2, -10 * time / 1) + 1)
+func exponentialOut(time: Double) -> Double {
+    time == 1 ? 1 : (-pow(2, -10 * time / 1) + 1)
 }
 
-func exponentialIn(time: Float) -> Float {
-    time == 0 ? 0 : powf(2, 10 * (time / 1 - 1)) - 1 * 0.001
+func exponentialIn(time: Double) -> Double {
+    time == 0 ? 0 : pow(2, 10 * (time / 1 - 1)) - 1 * 0.001
 }
 
-func exponentialInOut(time: Float) -> Float {
+func exponentialInOut(time: Double) -> Double {
     let time = time / 0.5
     if time < 1 {
-        return 0.5 * powf(2, 10 * (time - 1))
+        return 0.5 * pow(2, 10 * (time - 1))
     } else {
-        return 0.5 * (-powf(2, -10 * (time - 1)) + 2)
+        return 0.5 * (-pow(2, -10 * (time - 1)) + 2)
     }
 }
 
-func elasticIn(time: Float, period: Float) -> Float {
+func elasticIn(time: Double, period: Double) -> Double {
     if time == 0 || time == 1 {
         return time
     } else {
         let s = period / 4
         let rtime = time - 1
-        return -(powf(2, 10 * rtime) * sin((rtime - s) * .pi * 2 / period))
+        return -(pow(2, 10 * rtime) * sin((rtime - s) * .pi * 2 / period))
     }
 }
 
-func elasticOut(time: Float, period: Float) -> Float {
+func elasticOut(time: Double, period: Double) -> Double {
     if time == 0 || time == 1 {
         return time
     } else {
         let s = period / 4
-        return powf (2, -10 * time) * sin((time - s) * .pi * 2 / period) + 1
+        return pow (2, -10 * time) * sin((time - s) * .pi * 2 / period) + 1
     }
 }
 
-func elasticInOut(time: Float, period: Float) -> Float {
+func elasticInOut(time: Double, period: Double) -> Double {
     if time == 0 || time == 1 {
         return time
     } else {
@@ -116,9 +116,9 @@ func elasticInOut(time: Float, period: Float) -> Float {
         let s = period / 4
         
         if time < 0 {
-            return (-0.5 * powf(2, 10 * time) * sin((time - s) * .pi * 2 / period))
+            return (-0.5 * pow(2, 10 * time) * sin((time - s) * .pi * 2 / period))
         } else {
-            return powf(2, -10 * time) * sin((time - s) * .pi*2 / period) * 0.5 + 1
+            return pow(2, -10 * time) * sin((time - s) * .pi*2 / period) * 0.5 + 1
         }
     }
 }
